requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
biopython==1.81
python-multipart==0.0.6

